routes fully functional

+++ db/migrate/20230902191524_create_customers.rb
@@ -4,7 +4,7 @@ class CreateCustomers < ActiveRecord::Migration[7.0]
       t.string :full_name, null: false
       t.string :email, null: false
       t.string :password_digest, null: false
-      t.integer :mobile_number, null: false
+      t.bigint :mobile_number, null: false
       t.integer :age, null: false
 
       t.timestamps
